% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nr-text.R
\name{nr_text}
\alias{nr_text}
\title{Render strings to a \code{nativeRaster} image}
\usage{
nr_text(
  strings,
  font_files,
  font_idx = 0,
  color = NA,
  scale = 1,
  native_raster = TRUE,
  ...
)
}
\arguments{
\item{strings}{strings to shape}

\item{font_files}{path to font}

\item{font_idx}{Index of font within \code{font_file}. Default: 0}

\item{color}{single colour}

\item{scale}{size scale factor. Default: 1.0}

\item{native_raster}{logical. Return a native raster? default: TRUE.
If false, just returns a raster object}

\item{...}{further arguments passed to \code{textshaping::shape_text()}}
}
\value{
\code{nativeRaster}
}
\description{
Render strings to a \code{nativeRaster} image
}
\examples{
strings <- c('Ping', ' #RStats')
font_files <- c(
  system.file("fonts", "Montserrat", "static", "Montserrat-BoldItalic.ttf", package="grDevices"),
  system.file("fonts", "Montserrat", "static", "Montserrat-Medium.ttf"    , package="grDevices")
)
color <- c('red', 'blue')
txt <- nr_text(strings, font_files = font_files, color = color)
plot(txt, T)

}
